+++ src/Facades/Firebase.php
@@ -23,7 +23,7 @@ use Illuminate\Support\Facades\Facade;
  */
 final class Firebase extends Facade
 {
-    protected static function getFacadeAccessor()
+    protected static function getFacadeAccessor(): string
     {
         return 'firebase.manager';
     }