+++ src/ServiceProvider.php
@@ -10,7 +10,7 @@ use Laravel\Lumen\Application as Lumen;
 
 final class ServiceProvider extends \Illuminate\Support\ServiceProvider
 {
-    public function boot()
+    public function boot(): void
     {
         // @codeCoverageIgnoreStart
         if (!$this->app->runningInConsole()) {
@@ -27,7 +27,7 @@ final class ServiceProvider extends \Illuminate\Support\ServiceProvider
         ], 'config');
     }
 
-    public function register()
+    public function register(): void
     {
         // @codeCoverageIgnoreStart
         if ($this->app instanceof Lumen) {
@@ -43,48 +43,31 @@ final class ServiceProvider extends \Illuminate\Support\ServiceProvider
 
     private function registerComponents(): void
     {
-        $this->app->singleton(Firebase\Contract\Auth::class, static function (Container $app) {
-            return $app->make(FirebaseProjectManager::class)->project()->auth();
-        });
-        $this->app->alias(Firebase\Contract\Auth::class, Firebase\Auth::class);
+        $this->app->singleton(Firebase\Contract\Auth::class, static fn (Container $app) => $app->make(FirebaseProjectManager::class)->project()->auth());
         $this->app->alias(Firebase\Contract\Auth::class, 'firebase.auth');
 
-        $this->app->singleton(Firebase\Database::class, static function (Container $app) {
-            return $app->make(FirebaseProjectManager::class)->project()->database();
-        });
-        $this->app->alias(Firebase\Database::class, 'firebase.database');
+        $this->app->singleton(Firebase\Contract\Database::class, static fn (Container $app) => $app->make(FirebaseProjectManager::class)->project()->database());
+        $this->app->alias(Firebase\Contract\Database::class, 'firebase.database');
 
-        $this->app->singleton(Firebase\DynamicLinks::class, static function (Container $app) {
-            return $app->make(FirebaseProjectManager::class)->project()->dynamicLinks();
-        });
-        $this->app->alias(Firebase\DynamicLinks::class, 'firebase.dynamic_links');
+        $this->app->singleton(Firebase\Contract\DynamicLinks::class, static fn (Container $app) => $app->make(FirebaseProjectManager::class)->project()->dynamicLinks());
+        $this->app->alias(Firebase\Contract\DynamicLinks::class, 'firebase.dynamic_links');
 
-        $this->app->singleton(Firebase\Firestore::class, static function (Container $app) {
-            return $app->make(FirebaseProjectManager::class)->project()->firestore();
-        });
-        $this->app->alias(Firebase\Firestore::class, 'firebase.firestore');
+        $this->app->singleton(Firebase\Contract\Firestore::class, static fn (Container $app) => $app->make(FirebaseProjectManager::class)->project()->firestore());
+        $this->app->alias(Firebase\Contract\Firestore::class, 'firebase.firestore');
 
-        $this->app->singleton(Firebase\Messaging::class, static function (Container $app) {
-            return $app->make(FirebaseProjectManager::class)->project()->messaging();
-        });
-        $this->app->alias(Firebase\Messaging::class, 'firebase.messaging');
+        $this->app->singleton(Firebase\Contract\Messaging::class, static fn (Container $app) => $app->make(FirebaseProjectManager::class)->project()->messaging());
+        $this->app->alias(Firebase\Contract\Messaging::class, 'firebase.messaging');
 
-        $this->app->singleton(Firebase\RemoteConfig::class, static function (Container $app) {
-            return $app->make(FirebaseProjectManager::class)->project()->remoteConfig();
-        });
-        $this->app->alias(Firebase\RemoteConfig::class, 'firebase.remote_config');
+        $this->app->singleton(Firebase\Contract\RemoteConfig::class, static fn (Container $app) => $app->make(FirebaseProjectManager::class)->project()->remoteConfig());
+        $this->app->alias(Firebase\Contract\RemoteConfig::class, 'firebase.remote_config');
 
-        $this->app->singleton(Firebase\Storage::class, static function (Container $app) {
-            return $app->make(FirebaseProjectManager::class)->project()->storage();
-        });
-        $this->app->alias(Firebase\Storage::class, 'firebase.storage');
+        $this->app->singleton(Firebase\Contract\Storage::class, static fn (Container $app) => $app->make(FirebaseProjectManager::class)->project()->storage());
+        $this->app->alias(Firebase\Contract\Storage::class, 'firebase.storage');
     }
 
     private function registerManager(): void
     {
-        $this->app->singleton(FirebaseProjectManager::class, static function (Container $app) {
-            return new FirebaseProjectManager($app);
-        });
+        $this->app->singleton(FirebaseProjectManager::class, static fn (Container $app) => new FirebaseProjectManager($app));
         $this->app->alias(FirebaseProjectManager::class, 'firebase.manager');
     }
 }