+++ tests/ServiceProviderTest.php
@@ -18,12 +18,12 @@ final class ServiceProviderTest extends TestCase
     {
         $this->app->config->set('firebase.projects.app.credentials.file', \realpath(__DIR__.'/_fixtures/service_account.json'));
 
-        $this->assertInstanceOf(Firebase\Auth::class, $this->app->make(Firebase\Auth::class));
-        $this->assertInstanceOf(Firebase\Database::class, $this->app->make(Firebase\Database::class));
-        $this->assertInstanceOf(Firebase\DynamicLinks::class, $this->app->make(Firebase\DynamicLinks::class));
-        $this->assertInstanceOf(Firebase\Messaging::class, $this->app->make(Firebase\Messaging::class));
-        $this->assertInstanceOf(Firebase\RemoteConfig::class, $this->app->make(Firebase\RemoteConfig::class));
-        $this->assertInstanceOf(Firebase\Storage::class, $this->app->make(Firebase\Storage::class));
+        $this->assertInstanceOf(Firebase\Contract\Auth::class, $this->app->make(Firebase\Contract\Auth::class));
+        $this->assertInstanceOf(Firebase\Contract\Database::class, $this->app->make(Firebase\Contract\Database::class));
+        $this->assertInstanceOf(Firebase\Contract\DynamicLinks::class, $this->app->make(Firebase\Contract\DynamicLinks::class));
+        $this->assertInstanceOf(Firebase\Contract\Messaging::class, $this->app->make(Firebase\Contract\Messaging::class));
+        $this->assertInstanceOf(Firebase\Contract\RemoteConfig::class, $this->app->make(Firebase\Contract\RemoteConfig::class));
+        $this->assertInstanceOf(Firebase\Contract\Storage::class, $this->app->make(Firebase\Contract\Storage::class));
     }
 
     /**
@@ -34,6 +34,6 @@ final class ServiceProviderTest extends TestCase
         $this->expectException(\Throwable::class);
         $this->expectDeprecationMessageMatches('/unable/i');
 
-        $this->app->make(Firebase\Firestore::class);
+        $this->app->make(Firebase\Contract\Firestore::class);
     }
 }